adding start offset from end + bug fixing

+++ FileStream/ABCFile.h
@@ -9,10 +9,10 @@ class ABCFile : public IFile
 protected:
     ABCFile(IPrinter& _printer);
 public:
-    void setOffset(size_t _offset);
+    void setOffset(long _offset);
     void setLength(size_t _length);
 protected:
-    size_t    m_offset;
+    long      m_offset;
     size_t    m_length;
     IPrinter& m_printer;
 };
